# kvevent: refactor blocking buffer to use chunked linked list