--- a/pkg/ccl/changefeedccl/kvevent/blocking_buffer.go
+++ b/pkg/ccl/changefeedccl/kvevent/blocking_buffer.go
@@ -34,11 +34,11 @@ type blockingBuffer struct {
 
 	mu struct {
 		syncutil.Mutex
-		closed  bool             // True when buffer closed.
-		reason  error            // Reason buffer is closed.
-		drainCh chan struct{}    // Set when Drain request issued.
-		blocked bool             // Set when event is blocked, waiting to acquire quota.
-		queue   bufferEntryQueue // Queue of added events.
+		closed  bool                  // True when buffer closed.
+		reason  error                 // Reason buffer is closed.
+		drainCh chan struct{}         // Set when Drain request issued.
+		blocked bool                  // Set when event is blocked, waiting to acquire quota.
+		queue   bufferEventChunkQueue // Queue of added events.
 	}
 }
 
@@ -74,16 +74,16 @@ func NewMemBuffer(
 
 var _ Buffer = (*blockingBuffer)(nil)
 
-func (b *blockingBuffer) pop() (e *bufferEntry, err error) {
+func (b *blockingBuffer) pop() (e Event, ok bool, err error) {
 	b.mu.Lock()
 	defer b.mu.Unlock()
 	if b.mu.closed {
-		return nil, ErrBufferClosed{reason: b.mu.reason}
+		return Event{}, false, ErrBufferClosed{reason: b.mu.reason}
 	}
 
-	e = b.mu.queue.dequeue()
+	e, ok = b.mu.queue.dequeue()
 
-	if e == nil && b.mu.blocked {
+	if !ok && b.mu.blocked {
 		// Here, we know that we are blocked, waiting for memory; yet we have nothing queued up
 		// (and thus, no resources that could be released by draining the queue).
 		// This means that all the previously added entries have been read by the consumer,
@@ -93,7 +93,8 @@ func (b *blockingBuffer) pop() (e *bufferEntry, err error) {
 		// If the batching event consumer does not have periodic flush configured,
 		// we may never be able to make forward progress.
 		// So, we issue the flush request to the consumer to ensure that we release some memory.
-		e = newBufferEntry(Event{flush: true})
+		e = Event{flush: true}
+		ok = true
 		// Ensure we notify only once.
 		b.mu.blocked = false
 	}
@@ -102,7 +103,7 @@ func (b *blockingBuffer) pop() (e *bufferEntry, err error) {
 		close(b.mu.drainCh)
 		b.mu.drainCh = nil
 	}
-	return e, nil
+	return e, ok, nil
 }
 
 // notifyOutOfQuota is invoked by memQuota to notify blocking buffer that
@@ -125,16 +126,14 @@ func (b *blockingBuffer) notifyOutOfQuota() {
 // Get implements kvevent.Reader interface.
 func (b *blockingBuffer) Get(ctx context.Context) (ev Event, err error) {
 	for {
-		got, err := b.pop()
+		got, ok, err := b.pop()
 		if err != nil {
 			return Event{}, err
 		}
 
-		if got != nil {
+		if ok {
 			b.metrics.BufferEntriesOut.Inc(1)
-			e := got.e
-			bufferEntryPool.Put(got)
-			return e, nil
+			return got, nil
 		}
 
 		select {
@@ -160,15 +159,10 @@ func (b *blockingBuffer) ensureOpenedLocked(ctx context.Context) error {
 	return nil
 }
 
-func (b *blockingBuffer) enqueue(ctx context.Context, be *bufferEntry) (err error) {
+func (b *blockingBuffer) enqueue(ctx context.Context, e Event) (err error) {
 	// Enqueue message, and signal if anybody is waiting.
 	b.mu.Lock()
 	defer b.mu.Unlock()
-	defer func() {
-		if err != nil {
-			bufferEntryPool.Put(be)
-		}
-	}()
 
 	if err = b.ensureOpenedLocked(ctx); err != nil {
 		return err
@@ -176,7 +170,7 @@ func (b *blockingBuffer) enqueue(ctx context.Context, be *bufferEntry) (err erro
 
 	b.metrics.BufferEntriesIn.Inc(1)
 	b.mu.blocked = false
-	b.mu.queue.enqueue(be)
+	b.mu.queue.enqueue(e)
 
 	select {
 	case b.signalCh <- struct{}{}:
@@ -193,7 +187,7 @@ func (b *blockingBuffer) Add(ctx context.Context, e Event) error {
 
 	if e.alloc.ap != nil {
 		// Use allocation associated with the event itself.
-		return b.enqueue(ctx, newBufferEntry(e))
+		return b.enqueue(ctx, e)
 	}
 
 	// Acquire the quota first.
@@ -207,14 +201,20 @@ func (b *blockingBuffer) Add(ctx context.Context, e Event) error {
 		ap:      &b.qp,
 	}
 	e.bufferAddTimestamp = timeutil.Now()
-	be := newBufferEntry(e)
 
+	// An allocation must occur here because the event must fit the
+	// quotapool.Request interface. newBufferEntry and freeBufferEvent use a
+	// sync.Pool to mitigate the impact of this allocation. When
+	// blockingBuffer.Add is called with a single producer, the code below will
+	// just reuse the same pooled allocation.
+	be := newBufferEntry()
+	be.alloc = e.alloc
 	if err := b.qp.Acquire(ctx, be); err != nil {
-		bufferEntryPool.Put(be)
 		return err
 	}
+	freeBufferEvent(be)
 	b.metrics.BufferEntriesMemAcquired.Inc(alloc)
-	return b.enqueue(ctx, be)
+	return b.enqueue(ctx, e)
 }
 
 // tryDrain attempts to see if the buffer already empty.
@@ -284,11 +284,7 @@ func (b *blockingBuffer) CloseWithReason(ctx context.Context, reason error) erro
 	close(b.signalCh)
 
 	// Return all queued up entries to the buffer pool.
-	// Note: we do not need to release their resources since we are going to close
-	// bound account anyway.
-	for be := b.mu.queue.dequeue(); be != nil; be = b.mu.queue.dequeue() {
-		bufferEntryPool.Put(be)
-	}
+	b.mu.queue.purge()
 
 	return nil
 }
@@ -322,10 +318,9 @@ var _ quotapool.Resource = (*memQuota)(nil)
 // bufferEntry forms a linked list of elements in the buffer.
 // These entries are pooled to eliminate allocations.
 // bufferEntry also implements quotapool.Request interface for resource acquisition.
-type bufferEntry struct {
-	e    Event
-	next *bufferEntry // linked-list element
-}
+type bufferEntry Event
+
+var _ quotapool.Request = (*bufferEntry)(nil)
 
 var bufferEntryPool = sync.Pool{
 	New: func() interface{} {
@@ -333,14 +328,14 @@ var bufferEntryPool = sync.Pool{
 	},
 }
 
-func newBufferEntry(e Event) *bufferEntry {
-	be := bufferEntryPool.Get().(*bufferEntry)
-	be.e = e
-	be.next = nil
-	return be
+func newBufferEntry() *bufferEntry {
+	return bufferEntryPool.Get().(*bufferEntry)
 }
 
-var _ quotapool.Request = (*bufferEntry)(nil)
+func freeBufferEvent(c *bufferEntry) {
+	*c = bufferEntry{}
+	bufferEntryPool.Put(c)
+}
 
 // Acquire implements quotapool.Request interface.
 func (be *bufferEntry) Acquire(
@@ -348,7 +343,6 @@ func (be *bufferEntry) Acquire(
 ) (fulfilled bool, tryAgainAfter time.Duration) {
 	quota := resource.(*memQuota)
 	fulfilled, tryAgainAfter = be.acquireQuota(ctx, quota)
-
 	if !fulfilled {
 		quota.notifyOutOfQuota()
 	}
@@ -366,7 +360,7 @@ func (be *bufferEntry) acquireQuota(
 		quota.canAllocateBelow = 0
 	}
 
-	if err := quota.acc.Grow(ctx, be.e.alloc.bytes); err != nil {
+	if err := quota.acc.Grow(ctx, be.alloc.bytes); err != nil {
 		if quota.allocated == 0 {
 			// We've failed but there's nothing outstanding.  It seems that this request
 			// is doomed to fail forever. However, that's not the case since our memory
@@ -382,7 +376,7 @@ func (be *bufferEntry) acquireQuota(
 		return false, 0
 	}
 
-	quota.allocated += be.e.alloc.bytes
+	quota.allocated += be.alloc.bytes
 	quota.canAllocateBelow = 0
 	return true, 0
 }
@@ -390,35 +384,6 @@ func (be *bufferEntry) acquireQuota(
 // ShouldWait implements quotapool.Request interface.
 func (be *bufferEntry) ShouldWait() bool {
 	return true
-}
-
-// bufferEntryQueue is a queue implemented as a linked-list of bufferEntry.
-type bufferEntryQueue struct {
-	head, tail *bufferEntry
-}
-
-func (l *bufferEntryQueue) enqueue(be *bufferEntry) {
-	if l.tail == nil {
-		l.head, l.tail = be, be
-	} else {
-		l.tail.next = be
-		l.tail = be
-	}
-}
-
-func (l *bufferEntryQueue) empty() bool {
-	return l.head == nil
-}
-
-func (l *bufferEntryQueue) dequeue() *bufferEntry {
-	if l.head == nil {
-		return nil
-	}
-	ret := l.head
-	if l.head = l.head.next; l.head == nil {
-		l.tail = nil
-	}
-	return ret
 }
 
 type allocPool struct {